+++ packages/process/changelog.md
@@ -5,6 +5,7 @@
 * Refactor interpreters to use `PScoped` as the basis and specialize unparameterized ones.
   Don't provide loads of specializations for I/O types, rather offer interpreters that handle `ProcessOutput` and
   `ProcessInput` only.
+* Add constructors for `SysProcConf` that use `Path` and `Text`.
 
 # 0.9.0.0